+++ src/main/java/seedu/duke/Duke.java
@@ -1,13 +1,15 @@
 package seedu.duke;
 
+import seedu.duke.common.LogManager;
 import seedu.duke.common.Messages;
-import seedu.duke.contact.ContactManager;
+import seedu.duke.model.Model;
+import seedu.duke.model.contact.ContactManager;
 import seedu.duke.controller.ControlManager;
-import seedu.duke.event.EventManager;
-import seedu.duke.event.EventParameter;
+import seedu.duke.model.event.EventManager;
+import seedu.duke.model.event.EventParameter;
 import seedu.duke.controller.parser.CommandType;
-import seedu.duke.quiz.Quiz;
-import seedu.duke.quiz.QuizManager;
+import seedu.duke.model.quiz.Quiz;
+import seedu.duke.model.quiz.QuizManager;
 import seedu.duke.storage.QuizStorageManager;
 import seedu.duke.storage.EventStorageManager;
 import seedu.duke.ui.UserInterface;
