+++ src/main/java/seedu/duke/controller/command/AddClassCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.controller.command;
 
-import seedu.duke.Model;
+import seedu.duke.model.Model;
 import seedu.duke.exception.InvalidClassInputException;
 
 public class AddClassCommand extends Command {
