+++ src/main/java/seedu/duke/controller/command/DoneCcaCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.controller.command;
 
-import seedu.duke.Model;
+import seedu.duke.model.Model;
 
 public class DoneCcaCommand extends Command {
     public DoneCcaCommand(String userInput) {
