+++ src/main/java/seedu/duke/controller/command/DoneTuitionCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.controller.command;
 
-import seedu.duke.Model;
+import seedu.duke.model.Model;
 
 public class DoneTuitionCommand extends Command {
     public DoneTuitionCommand(String userInput) {
