+++ src/main/java/seedu/duke/model/event/cca/CcaManager.java
@@ -1,5 +1,7 @@
-package seedu.duke;
+package seedu.duke.model.event.cca;
 
+import seedu.duke.Event;
+import seedu.duke.common.LogManager;
 import seedu.duke.common.Messages;
 import seedu.duke.exception.CcaEmptyStringException;
 import seedu.duke.exception.CcaParamException;
