+++ src/main/java/seedu/duke/storage/EventStorageManager.java
@@ -1,14 +1,14 @@
 package seedu.duke.storage;
 
-import seedu.duke.Cca;
-import seedu.duke.Class;
+import seedu.duke.model.event.cca.Cca;
+import seedu.duke.model.event.classlesson.Class;
 import seedu.duke.Event;
-import seedu.duke.LogManager;
+import seedu.duke.common.LogManager;
 import seedu.duke.common.Messages;
-import seedu.duke.Test;
-import seedu.duke.Tuition;
+import seedu.duke.model.event.test.Test;
+import seedu.duke.model.event.tuition.Tuition;
 import seedu.duke.ui.UserInterface;
-import seedu.duke.event.EventParameter;
+import seedu.duke.model.event.EventParameter;
 
 import java.io.File;
 import java.io.IOException;
